Add netconf_capabilities plugin
