+++ nornir/plugins/tasks/networking/__init__.py
@@ -2,6 +2,7 @@ from .napalm_cli import napalm_cli
 from .napalm_configure import napalm_configure
 from .napalm_get import napalm_get
 from .napalm_validate import napalm_validate
+from .netconf_capabilities import netconf_capabilities
 from .netmiko_commit import netmiko_commit
 from .netmiko_file_transfer import netmiko_file_transfer
 from .netmiko_send_command import netmiko_send_command
@@ -14,6 +15,7 @@ __all__ = (
     "napalm_configure",
     "napalm_get",
     "napalm_validate",
+    "netconf_capabilities",
     "netmiko_commit",
     "netmiko_file_transfer",
     "netmiko_send_command",
